Experiment with using length instead of string
Things in favor of returning only the length:
- Saves 1 usize in memory per token.
- Can force usage of TextUnit.
- Is 'static.
- When manually writing all the code, it makes things cleaner.

Things in favor of returning the string reference:
- The memory usually isn't kept along for very long anyway, saving 1
  usize per token doesn't really matter.
- Keeping only the length is useless without also keeping the main
  string around. This is usable out of the box.
- Makes rowan-tools code less complicated with less steps. There's no
  need to have state -> length -> string when we can have state ->
  string.

I mostly ended up going with this because I got lifetime errors when
using &str instead of the 'static type SmolStr. And the reason to not
use SmolStr is of course to not force things upon the users.